Added more help files
