+++ src/main/java/net/sf/mzmine/util/maths/similarity/spectra/impl/composite/help/help.html
@@ -1,0 +1,27 @@
+<html>
+	<head>
+		<title>Spectral DB search - Spectral similarity - Composite weighted cosine</title>
+		<meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+		<link rel="stylesheet" type="text/css" href="/net/sf/mzmine/desktop/impl/helpsystem/HelpStyles.css">
+    </head>
+
+<body>
+
+<h1>Composite weighted cosine spectral similarity (identity)</h1>
+
+<h2>Description</h2>
+<p>
+Is used to determine the similarity between two spectra (usually library and query spectra). Especially useful for very reproducible generation of spectra (GC-EI-MS). Takes into account the relative intensities of neighbouring signals in the two spectra.
+</p>
+
+
+<h4>Method parameters</h4>
+<dl>
+<dt>Weights</dt>
+	<dd>Weights are applied to m/z and intensity: m/z<sup>a</sup>*Intensity<sup>b</sup></dd>
+<dt>Minimum cos similarity</dt>
+	<dd>Minimum cosine similarty for a match</dd>
+</dl>
+
+</body>
+</html>
