+++ src/main/java/net/sf/mzmine/util/maths/similarity/spectra/impl/cosine/help/help.html
@@ -1,0 +1,29 @@
+<html>
+	<head>
+		<title>Spectral DB search - Spectral similarity - Weighted cosine</title>
+		<meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+		<link rel="stylesheet" type="text/css" href="/net/sf/mzmine/desktop/impl/helpsystem/HelpStyles.css">
+    </head>
+
+<body>
+
+<h1>Weighted cosine spectral similarity</h1>
+
+<h2>Description</h2>
+<p>
+Is used to determine the similarity between two spectra (usually library and query spectra). 
+</p>
+
+
+<h4>Method parameters</h4>
+<dl>
+<dt>Weights</dt>
+	<dd>Weights are applied to m/z and intensity: m/z<sup>a</sup>*Intensity<sup>b</sup></dd>
+<dt>Minimum cos similarity</dt>
+	<dd>Minimum cosine similarty for a match</dd>
+<dt>Remove unmatched signals</dt>
+	<dd>(Sometimes called reversed cosine similarity) Before cosine calculation, remove all signals that were not found in both spectra (leads to higher cosine similarity, especially for spectra with low number of signals, but also to higher false-positive rate. </dd>
+</dl>
+
+</body>
+</html>
